# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Privet API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/voice/transcribe": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Transcribe Audio",
        "description": "Transcribe audio to text using STT.",
        "operationId": "transcribe_audio_api_v1_voice_transcribe_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_audio_api_v1_voice_transcribe_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/synthesize": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Synthesize Speech",
        "description": "Synthesize text to speech using TTS.",
        "operationId": "synthesize_speech_api_v1_voice_synthesize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SynthesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/process": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Process Voice Message",
        "description": "Complete voice processing pipeline: STT -> AI -> TTS.",
        "operationId": "process_voice_message_api_v1_voice_process_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_voice_message_api_v1_voice_process_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceProcessingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/voice/audio/{audio_id}": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Get Audio File",
        "description": "Download a generated audio file.",
        "operationId": "get_audio_file_api_v1_voice_audio__audio_id__get",
        "parameters": [
          {
            "name": "audio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Audio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_process_voice_message_api_v1_voice_process_post": {
        "properties": {
          "audio_file": {
            "type": "string",
            "format": "binary",
            "title": "Audio File"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "audio_file",
          "user_id"
        ],
        "title": "Body_process_voice_message_api_v1_voice_process_post"
      },
      "Body_transcribe_audio_api_v1_voice_transcribe_post": {
        "properties": {
          "audio_file": {
            "type": "string",
            "format": "binary",
            "title": "Audio File"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "audio_file"
        ],
        "title": "Body_transcribe_audio_api_v1_voice_transcribe_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SynthesisRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Text to synthesize"
          },
          "voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice",
            "description": "Voice ID or name"
          },
          "emotion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotion",
            "description": "Emotion for synthesis"
          },
          "speed": {
            "type": "number",
            "maximum": 4.0,
            "minimum": 0.25,
            "title": "Speed",
            "description": "Speech speed",
            "default": 1.0
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Language code"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SynthesisRequest",
        "description": "Request for text-to-speech synthesis."
      },
      "SynthesisResponse": {
        "properties": {
          "audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Url",
            "description": "URL to download audio"
          },
          "audio_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Base64",
            "description": "Base64 encoded audio"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds",
            "description": "Audio duration"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "TTS provider used"
          },
          "voice": {
            "type": "string",
            "title": "Voice",
            "description": "Voice used"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "voice"
        ],
        "title": "SynthesisResponse",
        "description": "Response from text-to-speech synthesis."
      },
...
      "TranscriptionResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Transcribed text"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Detected language"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "Transcription confidence"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds",
            "description": "Audio duration"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TranscriptionResponse",
        "description": "Response from voice transcription."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceProcessingResponse": {
        "properties": {
          "transcribed_text": {
            "type": "string",
            "title": "Transcribed Text",
            "description": "Transcribed text"
          },
          "response_text": {
            "type": "string",
            "title": "Response Text",
            "description": "AI response text"
          },
          "audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Url",
            "description": "Response audio URL"
          },
          "emotion_detected": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotion Detected",
            "description": "Detected emotion"
          },
          "facts_extracted": {
            "type": "integer",
            "title": "Facts Extracted",
            "description": "Number of facts extracted",
            "default": 0
          },
          "processing_time_ms": {
            "type": "integer",
            "title": "Processing Time Ms",
            "description": "Total processing time"
          }
        },
        "type": "object",
        "required": [
          "transcribed_text",
          "response_text",
          "processing_time_ms"
        ],
        "title": "VoiceProcessingResponse",
        "description": "Complete voice processing response."
      }
    }
  }
}