# privet_api/api/v1/endpoints/voice.py
"""Voice processing endpoints."""

from typing import Optional
from fastapi import APIRouter, Depends, HTTPException, status, File, UploadFile, Form
from fastapi.responses import Response
import base64
import io

from privet_api.api.deps import (
    get_audio_service
)
from privet_api.models.schemas.voice import (
    TranscriptionRequest,
    TranscriptionResponse,
    SynthesisRequest,
    SynthesisResponse,
    VoiceProcessingRequest,
    VoiceProcessingResponse,
    TTSProvider
)
from privet_api.core.config import settings

router = APIRouter()


@router.post("/transcribe", response_model=TranscriptionResponse)
async def transcribe_audio(
    audio_file: UploadFile = File(...),
    user_id: Optional[int] = Form(None),
    language: Optional[str] = Form(None)
) -> TranscriptionResponse:
    """Transcribe audio to text using STT."""
    
    # Validate file size
    if audio_file.size and audio_file.size > settings.upload_max_size_bytes:
        raise HTTPException(
            status_code=status.HTTP_413_REQUEST_ENTITY_TOO_LARGE,
            detail=f"File size exceeds maximum of {settings.upload_max_size_mb}MB"
        )
    
    # Validate file type
    allowed_types = ["audio/ogg", "audio/mpeg", "audio/wav", "audio/webm", "audio/mp4"]
    if audio_file.content_type not in allowed_types:
        raise HTTPException(
            status_code=status.HTTP_415_UNSUPPORTED_MEDIA_TYPE,
            detail=f"Unsupported audio format. Allowed: {', '.join(allowed_types)}"
        )
    
    # TODO: Integrate with actual audio service
    # audio_service = get_audio_service()
    # audio_bytes = await audio_file.read()
    # text = await audio_service.transcribe_voice(audio_bytes)
    
    # Mock response for now
    return TranscriptionResponse(
        text="This is a placeholder transcription.",
        language=language or "en",
        confidence=0.95,
        duration_seconds=3.5
    )


@router.post("/synthesize", response_model=SynthesisResponse)
async def synthesize_speech(
    request: SynthesisRequest
) -> SynthesisResponse:
    """Synthesize text to speech using TTS."""
    
    # TODO: Integrate with actual audio service
    # audio_service = get_audio_service()
    # audio_bytes = await audio_service.text_to_speech(
    #     text=request.text
    #     voice=request.voice
    #     emotion=request.emotion
    #     speed=request.speed
    # )
    
    # Mock response for now
    # Create a small silent audio file (WAV header + minimal data)
    mock_audio = b"RIFF$\x00\x00\x00WAVEfmt \x10\x00\x00\x00\x01\x00\x01\x00D\xac\x00\x00\x88X\x01\x00\x02\x00\x10\x00data\x00\x00\x00\x00"
    audio_base64 = base64.b64encode(mock_audio).decode()
    
    return SynthesisResponse(
        audio_base64=audio_base64,
        duration_seconds=1.0,
        provider=settings.tts_provider,
        voice=request.voice or settings.tts_voice
    )


@router.post("/process", response_model=VoiceProcessingResponse)
async def process_voice_message(
    audio_file: UploadFile = File(...),
    user_id: int = Form(...),
    language: Optional[str] = Form(None)
) -> VoiceProcessingResponse:
    """Complete voice processing pipeline: STT -> AI -> TTS."""
    
    import time
    start_time = time.time()
    
    # TODO: Full integration with services
    # 1. Transcribe audio
    # 2. Process through conversation manager
    # 3. Generate voice response
    # 4. Extract facts
    
    processing_time_ms = int((time.time() - start_time) * 1000)
    
    return VoiceProcessingResponse(
        transcribed_text="Placeholder transcription",
        response_text="Placeholder AI response",
        audio_url=None,
        emotion_detected="neutral",
        facts_extracted=0,
        processing_time_ms=processing_time_ms
    )


@router.get("/providers", response_model=list[TTSProvider])
async def get_tts_providers(
) -> list[TTSProvider]:
    """Get available TTS providers and their capabilities."""
    
    providers = [
        TTSProvider(
            name="openai",
            voices=[
                {"id": "alloy", "name": "Alloy"},
                {"id": "echo", "name": "Echo"},
                {"id": "fable", "name": "Fable"},
                {"id": "onyx", "name": "Onyx"},
                {"id": "nova", "name": "Nova"},
                {"id": "shimmer", "name": "Shimmer"}
            ],
            languages=["en", "es", "fr", "de", "it", "pt", "ru", "ja", "ko", "zh"],
            features={
                "emotion_support": False,
                "speed_control": True,
                "voice_cloning": False
            }
        ),
        TTSProvider(
            name="elevenlabs",
            voices=[
                {"id": "rachel", "name": "Rachel"},
                {"id": "domi", "name": "Domi"},
                {"id": "bella", "name": "Bella"}
            ],
            languages=["en", "es", "fr", "de", "it", "pt", "pl", "ru"],
            features={
                "emotion_support": True,
                "speed_control": True,
                "voice_cloning": True
            }
        )
    ]
    
    return providers


@router.get("/audio/{audio_id}")
async def get_audio_file(
    audio_id: str
) -> Response:
    """Download a generated audio file."""
    
    # TODO: Implement file storage and retrieval
    # For now, return a mock audio response
    
    mock_audio = b"RIFF$\x00\x00\x00WAVEfmt \x10\x00\x00\x00\x01\x00\x01\x00D\xac\x00\x00\x88X\x01\x00\x02\x00\x10\x00data\x00\x00\x00\x00"
    
    return Response(
        content=mock_audio,
        media_type="audio/wav",
        headers={
            "Content-Disposition": f"attachment; filename={audio_id}.wav"
        }
    )